todo 商品添加优化， 完成了商品下单列表接口

+++ src/main/java/com/lei/mall/model/entity/Item.java
@@ -49,7 +49,7 @@ public class Item implements Serializable {
      * 产生订单数量
      */
     @TableField(value = "order_count")
-    private Integer order_count;
+    private Long order_count;
 
     /**
      * 积分价格
